Use a single random-number generator
There is no need for multiple random-number generators, especially not
for the Rochester correction. Using a single seed allows to produce
results in a deterministic way, which is useful for validation of the
code.

The new generator is TRandom3, which is also better than TRandom used
for the Rochester correction before.