Pass Options to constructor of LooperMain
Normally would also have a static method in LooperMain to register
options, whose output would be fed into the constructor of Options. But
the definitions of the options will need to be reshuffled again when
analysis code is separated out of LooperMain. Leave the definitions in
the main program for the time being.