Move btagger::fill_eff inside LooperMain
This function will later be moved into the class for the main HZZ
analysis. The other functions in the namespace btagger will be put
inside a class that will implement the b tag reweighting. The old
function btagger::fill_eff doesn't belong there.

Drop the filling of histograms for b-tagging efficiencies from the NRB
analysis. Only the histograms produced in the main analysis are used.

Change indexing in the anticipation of the conversion to TTreeReader.