Add class to build electrons
Adapt loopers to use produced collections of electrons instead of the
former TLorentzVectorWithIndex.

This changes a few histograms produced by the main looper. All these
histograms involve jets. The reason for the difference is that jets are
now cleaned against all tight electrons, not just the two first ones.
Histograms filled after the lepton veto are not affected.