Add a class to build photons
For simplicity, only clean the collection of photons against (tight)
electrons but not muons. It is not expected that muons can be
misreconstructed as photons.

In the instrumental MET analysis, do not fill components of photon ID in
photon_evt::Fill_photonEvt. Doing so with the class-based descripion of
photons would require storing all these low-level details in the class.
But the corresponding histograms are not used anywhere in the analysis.
However, do not delete the relevant code for the moment in case I'm
missing something.