Add a class to build jets
The collection of jets in an event is represented by a vector of objects
of class Jet. Since this class includes the value of the b-tagging
discriminator and the jet flavour, code that deals with b-tagging is
updated accordingly.

Instead of having a separate collection of central jets to be used for
b-tagging, use the main collection and apply the selection |eta| < 2.5
as needed. In future this should be replaced by a method that identifies
taggable jets.

Implement jet ID in the analysis code instead of reading the precomputed
decision from Baobabs since this is done for the outer part of the
detector anyway. As a side effect, fix #9. This change in jet ID affects
histograms produced in the main analysis.