Apply EW corrections based on name of input file
In class EWCorrectionWeight, use the name of one of the files from the
input data set to guess the physics content of the sample. This is a
change from using the path to the catalog file for the same purpose,
which is a change mistakenly introduced in commit [1]. Using the path to
the catalog file is a bug because the catalog is renamed simply to
"theLocalData.txt" when running with a local copy of input files (see
script Tools/prepareAllJobs.py).

[1] 60eb8ab6e7c832596c5f80aa689b098fac26a35b