Correct propagation of muon corrections to ptmiss
For muons with pt > 200 GeV, their full four-momenta were used in the
type 1 corrections for ptmiss, instead of just the difference between
the corrected and original four-momenta. This had a drastic impact on mT
distributions in the dimuon channel, as reported in the beginning of
this talk [1]. The ptmiss has been affected by this problem since the
type 1 correction was introduced for the first time in commit [2].

In addition, lepton ID and trigger scale factors for the affected muons
were read using pt = 0, which lead to wrong values for the corresponding
weights.

[1] https://indico.cern.ch/event/812016/#165-some-issues-to-be-resolved
[2] 0a65a51dac82781262c76389484a47385b88ac10