Remove trigger selection in NRB analysis
This allows to eliminate flag runOnBaobabs_ in preparation to the
migration to NanoAOD. Dropping the trigger selection completely as
opposed to applying it unconditionally was proposed by Hanwen. The
trigger selection was already applied when bonzais were produced.
Dropping it also allows to remove the corresponding method from
LooperMain, which is useful as the trigger selection will be implemented
differently on NanoAOD. The main analysis also doesn't apply the trigger
selection in its main branch.