Avoid double reporting of errors in Options
No immediate error should be reported on an unknown option or failed
value check in Options::GetAs*. Instead, just the exception is enough.
If the exception is caught, this is not an error, and only a debug
message should be logged. If the exception is not caught, it will
automatically result in an error message.