Do not produce install rules
In installation rules generated by CMake by default, executables are
stripped of rpath. Because of this the executable installed into the
directory bin failed to start as it could not find the yaml-cpp library
within the standard search paths. At the same time it worked fine before
the installation (i.e. the executable in the build directory). This
problem was introduced in commit [1] when the yaml-cpp library was
added.

Change CMake configuration so that no install rules are generated.
Instead, the executable is put directly into the directory bin.

This is also needed in the view of a future separation into multiple
executables and one or more shared libraries. An installation for a
library implies that its header files are moved into a standard
location, and this is not what we intend to do.

[1] f91e4ef2e03143697328013a5555ddbdae7de122