Implement an interface to access generator weights
This addresses #21. In particular, the fluctuations in the photon+jets
samples should be fixed. No changes to VBF H->ZZ samples are expected.

In general, the nominal generator weights get scaled by a constant
factor, which will cancel out when normalizing by the integral of
totEventInBaobab_tot.

Checked the impact on the final mT distributions for POWHEG gg->ZZ as
produced in the main analysis. Did this for the nominal configuration
and up and down systematic variations in PDF, alpha_s, and ME scales.
There were no changes compared to [1].

[1] c5724bdd5bbaf54bf32cf7503c98ab1688bc3cc3