Set build type to Release by default
If cmake is run as described in the main README, the build type is not
specified, and as a result no optimization is applied. Change the
default build type to Release, so that -O3 is applied.

This decreases the CPU time by factor 2 (using POWHEG gg->ZZ as a
benchmark).

This fixes #19.