Class-based description of datasets
Add class DatasetInfo to read catalogue files. An instance of it and a
TTreeReader are aggregated in class Dataset, which is what is exposed to
consumers.

Validated this change by running the main analysis on two files from the
gg->ZZ sample. There were no changes in the produced histograms.