Dataset definition stems and fragments
In addition to using full YAML dataset definition files, add support for
incomplete fragments that include missing parts ("stems") from a
centrally provided file. This allows to keep the stems, which have to be
created by humans, in the repository. See [1] for details.

This completes the implementation of #24 in code, but analyses are still
using the old catalogues for the moment.

[1] https://gitlab.cern.ch/HZZ-IIHE/hzz2l2nu/wikis/dataset-definitions