In harvesting don't call hadd for single source
When merging ROOT files and when the list of sources happen to contain a
single file, it is presumably substantially faster to copy that file
instead of giving it to hadd. Such an optimization was originally
introduced by Hanwen when implementing MELA reweighting, but that change
was discarded.

Provide an interface for the hadd utility with a namesake function.
Clean up related code a bit.