Implement adapting job splitting
Choose the number of files to be processed per job based on the target
number of events per job. This relies on reading the total number of
selected events from the DDF. If that number is not available, roll back
to using the previous default of 25 files per job.

Datasets differ wildly in the number of events per file, so this change
helps to uniformize jobs in runtime.