Choose systematic variation based on dataset name
When submitting jobs and harvesting, find systematic uncertainties
affecting a given dataset based on the dataset name (as specified in the
DDF) as opposed to the name of the DDF file.

Ideally would support for glob-like or full regular expressions when
specifying the affected datasets, but this would require substantial
changes to prepareDataCards.py. Keep the matching simplistic as a
result.