Avoid a warning from harvester
When requesting local copy of input files (the default behaviour), DDFs
for them were placed in the same directory as PBS scripts. In
launchAnalysis.sh they were counted as jobs, which lead to a mismatch
between the number of "jobs" and the number of output files. Put the
DDFs into a dedicated directory to avoid the warning about failed jobs.