Allow out-of-source task dir in consuming scripts
In scripts that use results of harvesting, do not hard-code the path to
directories with input files with respect to $HZZ2L2NU_BASE. Instead,
accept the relevant directory as an argument.

Now it should be possible to run all steps in launchAnalysis.sh with a
task directory located not in the source tree.