Get sum of generator weights from DatasetInfo
Change the definition of event weights in simulation as described in
step 1 of the migration path in #14. Histogram "totEventInBaobab_tot"
now only contains a single bin whose content is 1. This fixes a small
problem with the computation of the baobab -> bonzai filtering
efficiency, which was mentioned in #14. This changes the normalization
of output histograms in simulation but not the final rescaled
histograms produced in dataMCcomparison.C and other scripts (except for
the small change due to the fix of the filtering efficiency).