Drop second signal in samples.h for main analysis
In commit [1], a second signal sample was added to the list in
samples.h. However, the corresponding dataset is only processed in the
MELA branch of the analysis but not the main one. This leads to a crash
in dataMCcomparison.C in the main branch. Only keep the second signal in
the MELA branch.
[1] cd600d5cfa5c060d8cf9a04bbd16924e6f818622