Switch to NanoAOD input samples
Only update samples for the main and NRB analysis. Skims for the
single-photon channel haven't been produced yet, and MELA reweighting is
not implemented.

The list of samples in the main analysis is now the same as in the NRB
one, except for the addition of MuonEG data in the latter case. This
fixes #17.