Add script to build templates for stat analysis
The script runs on files created by the DileptonTrees analysis. Compared
to the mT binning defined in SmartSelectionMonitor_hzz [1], drop first one
or two bins, which are nearly empty. In particular, drop bin
(150, 225) GeV in the 0j channel.
[1] https://gitlab.cern.ch/HZZ-IIHE/hzz2l2nu/blob/61dafd3aa50be77eb04f2cb8ecf37616ba6a1481/src/SmartSelectionMonitor_hzz.cc#L42-45

There seems to be a memory leak in the JIT compilation employed by
RDataFrame: the code for event filtering is compiled for each input file
(there are more that 200 of them) and never deleted. See this issue [2].
The total memory usage of the script is several GiB, which is tolerable
given that it runs only for about 15 minutes.
[2] https://sft.its.cern.ch/jira/browse/ROOT-9438