Move listSamplesToRun* files into config directory
Adjust the names of files with DDF stems to avoid confusion. Also
reflect the fact that they are shared across all years.

Partly propagate this change in names to the scripts in
WeightsAndDatadriven. These scripts are broken, though, because are not
aware of the year (which is included in names of files with lists of
samples for a while already).