Strip unnecessary logic from steering scripts
In prepare_jobs.py and harvest.py, avoid all configuration that is not
related to those scripts directly. They now only accept command line
arguments that directly affect their execution and, in case of
prepare_jobs.py, arguments that should be forwarded to runHZZanalysis.
Those arguments are forwarded as they are, and prepare_jobs.py doesn't
check them in any way. In particular, both scripts don't care which
specific analysis will be run. It's the responsibility of the caller to
ensure consistency between analysis type and other arguments.

Prefixes for output files are also supplied by the caller. In
launchAnalysis.sh, only provide non-trivial prefixes for analysis
"HZZdatadriven" because only in that configuration the same samples can
be processed with different selections in the same task.

Option --all-control-plots in runHZZanalysis is not set automatically
anymore. User can specify it if needed.

Related to #43.