Update external code to access JER factors
Relevant files are overwritten with ones from CMSSW_10_2_22. Very few
modifications are applied to them:
* Define macro STANDALONE in all these files.
* In one place switch to a built-in exception class from a CMS one.

There are some warnings during compilation. They will be addressed in a
separate commit.

This update adds support for individual components of the total JER
uncertainty, although they are not used in the analysis framework.