Implement pileup ID weights
Pileup ID efficiency in simulation is parameterized with an XGBoost
model. Due to limitations of the C API of the version of XGBoost
included in LCG, the computation of the efficiency requires memory
allocation and deallocation per jet. In general, this is not optimal,
but the computation takes around 10 us per jet, which is fast enough.