Apply EE noise mitigation in ptmiss
Implement it according to eq. 4 from [1]. The procedure is not exact
in the removal of the type 1 correction from (METFixEE2017 - MET) we
don't filter jets based on EM energy fraction and don't remove muons
from the jets; on the other hand, we don't consider jets failing the
jet ID.
[1] https://hypernews.cern.ch/HyperNews/CMS/get/met/710.html

This fixes #48.