Allow better approximation of standard type 1 corr
In the standard computation of the type 1 correction to missing pt, jets
with a high EM energy fractions are skipped and constituent muons are
removed from jets (#49). Allow approximating this procedure.

The procedure is approximated rather than implementing it exactly
because the EM energy fractions are not saved in NanoAOD for soft jets
and the muon subtraction is done by rescaling of the jet momentum (i.e.
its direction cannot be changed). Also, jets overlapping with leptons
(or photons -- in the corresponding analysis) are always skipped from
the type 1 correction.

This procedure is switched off by default, except for when the
precomputed type 1 correction is undone when applying the EE noise
mitigation in 2017.