Implement PDF variations
Individual PDF variations are combined on per-event basis. With Hessian
variations, they are added up in quadrature; with MC replicas, the
standard deviation of individual weights in each event is used.

Also implement alpha_s variation. When needed, it is rescaled to the
recommended variation of +-1.5e-3. This fixes shears#37.